[ADD] laravel ide helper for development, pa command for development only

+++ routes/web.php
@@ -4,12 +4,14 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\Facades\Artisan;
 
-Route::get('config-cache', function () {
-  Artisan::call('migrate:fresh --seed');
-  Artisan::call('cache:clear');
-  Artisan::call('config:clear');
-  Artisan::call('config:cache');
-});
+if (config('app.env') == 'local') {
+  Route::get('config-cache', function () {
+    Artisan::call('migrate:fresh --seed');
+    Artisan::call('cache:clear');
+    Artisan::call('config:clear');
+    Artisan::call('config:cache');
+  });
+}
 
 Route::get('receipt', function () {
   $order = App\Models\Cart::firstOrFail();
